Allow setting PREFIX during configure
This proposal should add the ability to use $$PREFIX to determine
where should we install the files, not hardcoded /usr/*.

After this commit, packagers need the following instructions to
build and install into /usr:

* qmake PREFIX=/usr .
* make
* sudo make install

Unfortunately this would break custom packaging_scripts **a lot**,
but this should be a cleaner approach to install files, either
for distribution maintainers or for upstream authors.

Also there are hardcoded detections for /usr/share/... in source
code and plugins. The impact of using a different PREFIX is yet
to be sorted out.